Update TradeEveryNSec.mq4
Added a new input parameter MinAccountValueToCloseTrades set to 1000.0.
In the CheckAndCloseProfitableTrades() function, we now check the account equity before closing a profitable trade. If the account equity is less than 1000.0 and negative, we will not close the trade, even if the profit exceeds the TakeProfit value.

This way, the EA will only close profitable trades when the account equity is either above the MinAccountValueToCloseTrades threshold or at least 0.0 (the account is not in a negative balance). This should help protect your account from closing profitable trades when the balance is in a drawdown.